+++ colourettu/__init__.py
@@ -11,7 +11,34 @@ of colours.
 
 from __future__ import absolute_import
 
+# METADATA
+
+__title__ = "colourettu"
+__description__ = "Colourettu is a collection of colour related functions."
+__url__ = "http://minchin.ca/colourettu/"
+__author__ = "William Minchin"
+__email__ = "w_minchin@hotmail.com"
+__license__ = "MIT License"
+__copyright__ = "Copyright (c) 2014-16 William Minchin"
+
+"""
+This project uses the Semantic Versioning scheme in conjunction with PEP 0440:
+
+    <http://semver.org/>
+    <https://www.python.org/dev/peps/pep-0440>
+
+Major versions introduce significant changes to the API, and backwards
+compatibility is not guaranteed. Minor versions are for new features and other
+backwards-compatible changes to the API. Patch versions are for bug fixes and
+internal code changes that do not affect the API.
+
+Version 0.x should be considered a development version with an unstable API,
+and backwards compatibility is not guaranteed for minor versions.
+"""
+__version__ = "1.1.1-dev.0"
+
+
+# Package Implementation
+
 from ._colour import *
 from ._palette import *
-
-from .__version__ import __version__